--- CineHive/src/main/java/com/example/CineHive/service/creditService/movie/NowPlayingMovieService.java
@@ -1,60 +1,0 @@
-package com.example.CineHive.service.creditService.movie;
-
-import com.example.CineHive.dto.video.movie.NowPlayingMovieDto;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.client.WebClient;
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-public class NowPlayingMovieService {
-
-    @Value("${tmdb.api.key}")
-    private String apiKey;
-
-    private final WebClient webClient;
-    private final ObjectMapper objectMapper;
-
-
-    public NowPlayingMovieService(WebClient.Builder webClientBuilder, ObjectMapper objectMapper) {
-        this.webClient = webClientBuilder.baseUrl("https://api.themoviedb.org/3").build();
-        this.objectMapper = objectMapper;
-    }
-
-    public List<NowPlayingMovieDto> getNowPlayingMovies(Pageable pageable) {
-        String response = webClient.get()
-                .uri("https://api.themoviedb.org/3/movie/now_playing?language=ko&page=" + (pageable.getPageNumber() + 1) + "&api_key=" + apiKey)
-                .header("Accept", "application/json")
-                .retrieve()
-                .bodyToMono(String.class)
-                .block();
-
-        List<NowPlayingMovieDto> moviePosters = new ArrayList<>();
-        if (response != null) {
-            try {
-                JsonNode rootNode = objectMapper.readTree(response);
-                JsonNode moviesNode = rootNode.path("results");
-
-                for (JsonNode movieNode : moviesNode) {
-                    Long movieId = movieNode.get("id").asLong();
-                    String posterPath = movieNode.get("poster_path").asText();
-
-
-                    NowPlayingMovieDto nowPlayingMovieDto = new NowPlayingMovieDto(movieId, posterPath);
-                    moviePosters.add(nowPlayingMovieDto);
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        } else {
-            System.out.println("응답이 없습니다.");
-        }
-        return moviePosters;
-    }
-
-
-}
